test/web-platform-tests/to-upstream/css/cssom-view/HTMLElement-offsetLeft.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTMLElement.prototype.offsetLeft</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-htmlelement-offsetleft">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_equals(el.offsetLeft, 0);

}, "offsetLeft should return 0 for an element disconnected from a document");
</script>
